contests/abc059/b/main_test.go
package main

import (
	"fmt"
	"os"
	"path/filepath"
	"testing"
)

func TestSolution(t *testing.T) {
	type in struct {
		a, b string
	}
	tests := []struct {
		in   in
		want string
	}{
		{in: in{a: "36", b: "24"}, want: "GREATER"},
		{in: in{a: "850", b: "3777"}, want: "LESS"},
		{in: in{a: "9720246", b: "22516266"}, want: "LESS"},
	}
	for i, tt := range tests {
		t.Run(fmt.Sprint(i), func(t *testing.T) {
			got := solution(tt.in.a, tt.in.b)
			if got != tt.want {
				t.Fatalf("in: %v got: %v want: %v", tt.in, got, tt.want)
			}
		})
	}
}

func Example_main() {
	fd, _ := os.Open(filepath.Join("testdata", "input.txt"))
	orgStdin := os.Stdin
	os.Stdin = fd
	defer func() {
		os.Stdin = orgStdin
		fd.Close()
	}()

	main()

	// Output: LESS
}
